model-microservice/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python-headless
numpy
requests
pydantic
python-multipart
httpx
python-dotenv